requirements.txt
chardet==5.2.0
javalang==0.13.0
tqdm==4.67.1
pytest==8.4.2
zai-sdk==0.0.4